requirements.txt
Django==4.0.3
django-filter==21.1
djangorestframework==3.13.1
psycopg2-binary==2.9.3
PyJWT==2.3.0

